Update ctlr command line to 6.7: pass json encode args only
Login shells of admin and user users are set to sh.php, so they cannot
drop to a shell command line now. And we pass only the args to sh.php,
as a json encoded string. There is no ctlr shell script anymore, and
admin and user users do not have home folders now.